+++ R/io.R
@@ -185,10 +185,20 @@ grep_sub = function(pattern, replacement, x, ...) {
 #'   \code{\link{url_filename}()}.
 #' @param ... Other arguments to be passed to \code{\link{download.file}()}
 #'   (except \code{method}).
+#' @note To allow downloading large files, the \code{timeout} option in
+#'   \code{\link{options}()} will be temporarily set to one hour (3600 seconds)
+#'   inside this function when this option has the default value of 60 seconds.
+#'   If you want a different \code{timeout} value, you may set it via
+#'   \code{options(timeout = N)}, where \code{N} is the number of seconds (not
+#'   60).
 #' @return The integer code \code{0} for success, or an error if none of the
 #'   methods work.
 #' @export
 download_file = function(url, output = url_filename(url), ...) {
+  if (getOption('timeout') == 60L) {
+    opts = options(timeout = 3600)  # one hour
+    on.exit(options(opts), add = TRUE)
+  }
   download = function(method = 'auto') download.file(url, output, ..., method = method)
   for (method in c(if (is_windows()) 'wininet', 'libcurl', 'auto')) {
     if (!inherits(try_silent(res <- download(method = method)), 'try-error') && res == 0)