+++ NEWS.md
@@ -6,6 +6,10 @@
 
 - Added a new function `retry()` to retry calling a function for a number of times in case of errors.
 
+## MINOR CHANGES
+
+- The `timeout` option in `options()` is set to 3600 seconds when it takes the default value of 60 seconds, which may not be enough for downloading large files (thanks, @matthewgson, yihui/tinytex#286).
+
 # CHANGES IN xfun VERSION 0.21
 
 ## NEW FEATURES